+++ browser/misc_metrics/BUILD.gn
@@ -142,7 +142,7 @@ source_set("unit_tests") {
     "//brave/browser/misc_metrics:misc_metrics_impl",
     "//brave/components/brave_rewards/core",
     "//brave/components/misc_metrics",
-    "//brave/components/p3a_utils",
+    "//brave/components/p3a_utils:test_support",
     "//chrome/browser",
     "//chrome/browser/bookmarks",
     "//chrome/browser/content_settings:content_settings_factory",
