+++ components/p3a_utils/event_relay.h
@@ -13,8 +13,6 @@
 #include "base/no_destructor.h"
 #include "base/observer_list.h"
 #include "base/observer_list_types.h"
-#include "base/scoped_observation.h"
-#include "third_party/abseil-cpp/absl/container/flat_hash_map.h"
 
 namespace p3a_utils {
 
@@ -46,27 +44,6 @@ class EventRelay {
   ~EventRelay();
 
   base::ObserverList<Observer> observers_;
-};
-
-// Test helper that observes EventRelay and caches custom attribute values.
-class TestEventRelayObserver : public EventRelay::Observer {
- public:
-  TestEventRelayObserver();
-  ~TestEventRelayObserver() override;
-
-  // Returns the cached value for |attribute_name|, or std::nullopt if the
-  // attribute was cleared or never set.
-  std::optional<std::string> GetCustomAttribute(
-      std::string_view attribute_name) const;
-
-  // EventRelay::Observer
-  void OnCustomAttributeSet(
-      std::string_view attribute_name,
-      std::optional<std::string_view> attribute_value) override;
-
- private:
-  absl::flat_hash_map<std::string, std::optional<std::string>> attributes_;
-  base::ScopedObservation<EventRelay, EventRelay::Observer> observation_{this};
 };
 
 }  // namespace p3a_utils
